try to fix type bug

+++ code/train_mcnet_3d.py
@@ -129,7 +129,7 @@ if __name__ == "__main__":
         for i_batch, sampled_batch in enumerate(trainloader):
             
             volume_batch, label_batch = sampled_batch['image'], sampled_batch['label']
-            volume_batch, label_batch = volume_batch.cuda(), label_batch.cuda()
+            volume_batch, label_batch = volume_batch.cuda(), label_batch.cuda().float()
 
             model.train()
             outputs = model(volume_batch)
